more correctly select best path

+++ src/navigation/navigation.cc
@@ -382,7 +382,7 @@ void Navigation::_update_global_path() {
   if (_rrt.IsFindingPath()) {
     // time path update
     // auto now = std::chrono::high_resolution_clock::now();
-    size_t iterations = 0;
+    int iterations = 0;
     _rrt.FindPath(_path, iterations);
     // auto dt = std::chrono::high_resolution_clock::now() - now;
     // auto us = std::chrono::duration_cast<std::chrono::microseconds>(dt).count();
